Refactor training script and update documentation for T4 optimization
- Renamed `train_t4.py` to `train.py` to streamline the training process for single T4 GPU.
- Updated `README.md` to reflect the new training command and clarify the T4-specific training setup.
- Removed the deprecated `train_t4.py` file and adjusted references in the codebase to point to the new `train.py`.
- Enhanced error messages in `inference.py` to guide users on the correct training script usage.
