+++ core/train.py
@@ -1,22 +1,12 @@
 #!/usr/bin/env python3
 """
-Main training script for blueberry-llm.
-
-This script provides a clean entry point for training GPU-adaptive LLM models
-with automatic optimization based on hardware capabilities.
-
-Usage:
-    python train.py                    # Use default MoE configuration
-    python train.py --config dev       # Use development configuration
-    python train.py --config rtx5090   # Use RTX 5090 optimized configuration
+T4-optimized training for Blueberry LLM
+Just run: python train.py
 """
 
 import os
 import sys
-import argparse
-import torch
-from torch.utils.data import DataLoader, random_split
-import time
+os.environ["TOKENIZERS_PARALLELISM"] = "false"
 
 # Add parent directory to path for imports when running directly
 if __name__ == "__main__":
@@ -24,257 +14,150 @@ if __name__ == "__main__":
     parent_dir = os.path.dirname(current_dir)
     sys.path.insert(0, parent_dir)
 
-# Import our modular components
-from configs import AdaptiveMoEModelConfig, get_rtx4090_config, get_rtx5090_config, get_development_config
-from data import load_and_cache_data, TextTokenDataset
-from models import AdaptiveMoEMinimalLLM, create_model
-from training import train_model, validate_training_setup
-from system import print_system_info, SYSTEM_CONFIG
+import torch
+import argparse
+from torch.utils.data import DataLoader, random_split
+from core.t4_config import t4_configure
+from legacy.llm import train_moe_model, load_and_cache_data, TextTokenDataset
 
+def auto_launch_distributed():
+    """Single T4 GPU - no distributed training needed"""
+    return False
 
 def parse_arguments():
     """Parse command line arguments."""
     parser = argparse.ArgumentParser(
-        description="Train GPU-adaptive LLM models",
+        description="T4-optimized training for Blueberry LLM",
         formatter_class=argparse.ArgumentDefaultsHelpFormatter
     )
     
-    # Configuration options
-    parser.add_argument(
-        "--config", 
-        type=str, 
-        default="default",
-        choices=["default", "dev", "rtx4090", "rtx5090"],
-        help="Configuration preset to use"
-    )
-    
-    # Model options
-    parser.add_argument("--model-type", type=str, default="moe", choices=["moe", "standard"],
-                       help="Type of model to train")
-    parser.add_argument("--d-model", type=int, help="Model dimension")
-    parser.add_argument("--n-layers", type=int, help="Number of layers")
-    parser.add_argument("--n-heads", type=int, help="Number of attention heads")
-    parser.add_argument("--num-experts", type=int, help="Number of experts (MoE only)")
-    
-    # Training options
-    parser.add_argument("--max-steps", type=int, help="Maximum training steps")
-    parser.add_argument("--batch-size", type=int, help="Batch size")
-    parser.add_argument("--lr", type=float, help="Learning rate")
-    parser.add_argument("--no-fp8", action="store_true", help="Disable FP8 acceleration")
-    parser.add_argument("--no-amp", action="store_true", help="Disable automatic mixed precision")
-    
-    # Data options
-    parser.add_argument("--num-documents", type=int, help="Number of documents to load")
-    parser.add_argument("--max-tokens", type=int, help="Maximum number of tokens")
-    parser.add_argument("--max-seq-len", type=int, help="Maximum sequence length")
-    
-    # System options
-    parser.add_argument("--device", type=str, help="Device to use (cuda/cpu)")
-    parser.add_argument("--seed", type=int, default=42, help="Random seed")
-    parser.add_argument("--validate-setup", action="store_true", help="Validate setup before training")
-    parser.add_argument("--no-megatron", action="store_true", help="Force native backend (disable Megatron)")
-    parser.add_argument("--use-megatron", action="store_true", help="Force Megatron backend (enable Megatron)")
+    # Single T4 GPU - no Megatron options needed
     
     return parser.parse_args()
 
-
-def get_config(args):
-    """Get configuration based on arguments."""
-    # Base configuration
-    if args.config == "dev":
-        config = get_development_config()
-    elif args.config == "rtx4090":
-        config = get_rtx4090_config()
-    elif args.config == "rtx5090":
-        config = get_rtx5090_config()
-    else:
-        config = AdaptiveMoEModelConfig()
-    
-    # Override with command line arguments
-    if args.d_model is not None:
-        config.d_model = args.d_model
-    if args.n_layers is not None:
-        config.n_layers = args.n_layers
-    if args.n_heads is not None:
-        config.n_heads = args.n_heads
-    if args.num_experts is not None:
-        config.num_experts = args.num_experts
-    if args.max_steps is not None:
-        config.max_steps = args.max_steps
-    if args.batch_size is not None:
-        config.batch_size = args.batch_size
-    if args.lr is not None:
-        config.muon_lr = args.lr
-    if args.num_documents is not None:
-        config.num_documents = args.num_documents
-    if args.max_tokens is not None:
-        config.max_tokens = args.max_tokens
-    if args.max_seq_len is not None:
-        config.max_seq_len = args.max_seq_len
-    
-    # Disable features if requested
-    if args.no_fp8:
-        config.use_fp8 = False
-    if args.no_amp:
-        config.use_amp = False
-    if args.no_megatron:
-        config.use_megatron = False
-    if args.use_megatron:
-        config.use_megatron = True
-    
-    # Re-run post_init to update dependent values
-    config.__post_init__()
-    
-    return config
-
-
-def setup_data(config: AdaptiveMoEModelConfig):
-    """Setup data loaders."""
-    print("\n📚 Setting up data...")
-    
-    # Load and cache data
-    texts, tokenizer, tokens = load_and_cache_data(config)
-    
-    # Create dataset
-    dataset = TextTokenDataset(tokens, config.max_seq_len)
-    
-    # Train/validation split
-    val_size = len(dataset) // 10
-    train_size = len(dataset) - val_size
-    
-    generator = torch.Generator().manual_seed(42)
-    train_dataset, val_dataset = random_split(dataset, [train_size, val_size], generator=generator)
-    
-    # Create data loaders
-    train_loader = DataLoader(
-        train_dataset, 
-        batch_size=config.batch_size, 
-        shuffle=True, 
-        num_workers=2,
-        pin_memory=torch.cuda.is_available()
-    )
-    
-    val_loader = DataLoader(
-        val_dataset, 
-        batch_size=config.batch_size, 
-        shuffle=False, 
-        num_workers=2,
-        pin_memory=torch.cuda.is_available()
-    )
-    
-    print(f"✅ Dataset: {len(train_dataset):,} train, {len(val_dataset):,} val samples")
-    print(f"✅ Vocab size: {config.vocab_size:,}")
-    
-    return train_loader, val_loader, tokenizer
-
-
-def setup_model(config: AdaptiveMoEModelConfig, model_type: str):
-    """Setup model."""
-    print(f"\n🤖 Setting up {model_type} model...")
-    
-    # Create model
-    model = create_model(config, model_type)
-    
-    # Move to appropriate device
-    if torch.cuda.is_available():
-        model = model.cuda()
-        print(f"✅ Model moved to GPU")
-    else:
-        print(f"⚠️ Using CPU (CUDA not available)")
-    
-    return model
-
-
-def set_seed(seed: int):
-    """Set all random seeds for reproducibility."""
-    import random
-    import numpy as np
-    
-    random.seed(seed)
-    np.random.seed(seed)
-    torch.manual_seed(seed)
-    torch.cuda.manual_seed_all(seed)
-    torch.backends.cudnn.deterministic = True
-    torch.backends.cudnn.benchmark = False
-    print(f"🌱 Set all seeds to {seed}")
-
-
 def main():
-    """Main training function."""
+    print("🫐 Starting Blueberry LLM T4-Training")
+    
     # Parse arguments
     args = parse_arguments()
     
-    # Set random seed
-    set_seed(args.seed)
+    # Auto-launch with torchrun if needed
+    auto_launch_distributed()
     
-    # Print header
-    print("🚀 Blueberry LLM Training")
-    print("=" * 60)
+    # T4-configure everything
+    configurator = t4_configure()
     
-    # Print system information
+    # Single T4 GPU - no Megatron support
+    print("🚀 Single T4 GPU training - Megatron disabled")
+    
+    configurator.print_config()
+    
+    # Print detailed GPU system information
+    print("\n🔍 Detailed GPU System Information:")
+    print("=" * 50)
+    from system import print_system_info
     print_system_info()
+    print("=" * 50)
     
-    # Get configuration
-    config = get_config(args)
+    # Single T4 GPU - no distributed training needed
+    print("🚀 Single T4 GPU training mode")
     
-    # Setup data
-    train_loader, val_loader, tokenizer = setup_data(config)
+    # Get model configuration
+    model_config = configurator.get_model_config()
     
-    # Setup model
-    model = setup_model(config, args.model_type)
+    # Auto-size dataset for T4 GPU
+    model_config.num_documents = 2000
+    model_config.max_tokens = 200000
     
-    # Validate setup if requested
-    if args.validate_setup:
-        print("\n🔍 Validating training setup...")
-        if not validate_training_setup(model, train_loader, val_loader, config):
-            print("❌ Setup validation failed. Exiting.")
-            return
-        print("✅ Setup validation successful!")
+    print(f"\n📊 Loading {model_config.num_documents} documents, {model_config.max_tokens:,} tokens...")
     
-    # Setup device
-    if args.device:
-        device = torch.device(args.device)
-    else:
+    # Load data
+    texts, tokenizer, tokens = load_and_cache_data(model_config)
+    dataset = TextTokenDataset(tokens, model_config.max_seq_len)
+    
+    # Train/val split
+    val_size = len(dataset) // 10
+    train_size = len(dataset) - val_size
+    train_dataset, val_dataset = random_split(
+        dataset, [train_size, val_size], 
+        generator=torch.Generator().manual_seed(42)
+    )
+    
+    # Create data loaders for single T4 GPU
+    train_loader = DataLoader(
+        train_dataset, 
+        batch_size=model_config.batch_size, 
+        shuffle=True, 
+        num_workers=2
+    )
+    val_loader = DataLoader(
+        val_dataset, 
+        batch_size=model_config.batch_size, 
+        shuffle=False, 
+        num_workers=2
+    )
+    
+    print(f"   Dataset: {len(train_dataset)} train, {len(val_dataset)} val samples")
+    
+    # Train the model
+    print("\n🚀 Starting training...")
+    
+    # Use Megatron-enabled training if requested
+    if configurator.config.use_megatron:
+        print("🚀 Using Megatron-enabled training pipeline...")
+        from models import create_model
+        from training import train_model
+        from configs import AdaptiveMoEModelConfig
+        
+        # Convert legacy config to new config format
+        adaptive_config = AdaptiveMoEModelConfig(
+            d_model=model_config.d_model,
+            n_heads=model_config.n_heads,
+            n_layers=model_config.n_layers,
+            d_ff=model_config.d_ff,
+            batch_size=model_config.batch_size,
+            max_steps=model_config.max_steps,
+            gradient_accumulation_steps=model_config.gradient_accumulation_steps,
+            muon_lr=model_config.muon_lr,
+            max_seq_len=model_config.max_seq_len,
+            num_experts=model_config.num_experts,
+            use_amp=model_config.use_amp,
+            vocab_size=model_config.vocab_size  # Add vocab_size from tokenizer
+        )
+        
+        # Create model optimized for T4
+        model = create_model(adaptive_config, "moe")
+        
+        # Move to device
         device = torch.device('cuda' if torch.cuda.is_available() else 'cpu')
-    
-    print(f"\n🎯 Starting training on {device}")
-    print(f"📋 Configuration: {args.config}")
-    print(f"🤖 Model type: {args.model_type}")
-    
-    # Start training
-    start_time = time.time()
-    
-    try:
+        model = model.to(device)
+        
+        # Train with new pipeline
         model, final_metrics = train_model(
             model=model,
             train_loader=train_loader,
             val_loader=val_loader,
-            config=config,
+            config=adaptive_config,
             device=device
         )
-        
-        training_time = time.time() - start_time
-        
-        # Print final results
-        print("\n" + "=" * 60)
-        print("🎉 Training completed successfully!")
-        print(f"⏱️ Total training time: {training_time/60:.1f} minutes")
-        print("\n📊 Final Results:")
-        print(f"   Validation Loss: {final_metrics['val_loss']:.4f}")
-        print(f"   Validation Accuracy: {final_metrics['val_accuracy']:.4f}")
-        print(f"   Validation Perplexity: {final_metrics['val_perplexity']:.2f}")
-        
-        if 'mfu' in final_metrics:
-            print(f"   Model FLOPs Utilization: {final_metrics['mfu']*100:.1f}%")
-        
-        print("=" * 60)
-        
-    except KeyboardInterrupt:
-        print("\n⚠️ Training interrupted by user")
-    except Exception as e:
-        print(f"\n❌ Training failed with error: {e}")
-        raise
-
+    else:
+        # Use legacy training pipeline
+        model, final_metrics = train_moe_model(model_config, train_loader, val_loader)
+    
+    # Save results
+    print("\n💾 Saving model...")
+    torch.save({
+        'model_state_dict': model.state_dict(),
+        'config': model_config,
+        't4_config': configurator.config,
+        'tokenizer': tokenizer,
+        'final_metrics': final_metrics
+    }, 'blueberry_model.pt')
+    
+    print("✅ Training complete!")
+    print(f"   Final validation loss: {final_metrics['val_loss']:.4f}")
+    print(f"   Final validation accuracy: {final_metrics['val_accuracy']:.4f}")
+    print(f"   Model saved as: blueberry_model.pt")
 
 if __name__ == "__main__":
     main()
